fix: scripts folder structure
- 스크립트 폴더 내 구조 변경
GameObject의 Script 따로 분류
그 외 기타 engine 변경에 따른 변경
