--- Source/Scripts/CMissileScript.h
@@ -1,22 +1,0 @@
-﻿#pragma once
-#include <Engine/CScript.h>
-
-
-class CMissileScript :
-    public CScript
-{
-    Vec3 m_Velocity;
-
-public:
-    void SetVelocity(Vec3 _Velocity) { m_Velocity = _Velocity; }
-    Vec3 GetVelocity() { return m_Velocity; }
-
-    void Begin() override;
-    void Tick() override;
-    void SaveComponent(FILE* _File) override;
-    void LoadComponent(FILE* _File) override;
-
-    CLONE(CMissileScript);
-    CMissileScript();
-    ~CMissileScript() override;
-};
