--- Source/Scripts/CScriptMgr.h
@@ -1,25 +1,0 @@
-﻿#pragma once
-
-#include <intsafe.h>
-#include <vector>
-#include <string>
-
-enum SCRIPT_TYPE
-{
-    CAMERASCRIPT,
-    MISSILESCRIPT,
-    PLAYERSCRIPT,
-};
-
-using namespace std;
-
-class CScript;
-
-class CScriptMgr
-{
-public:
-    static void GetScriptInfo(vector<wstring>& _vec);
-    static CScript* GetScript(const wstring& _strScriptName);
-    static CScript* GetScript(UINT _iScriptType);
-    static const wchar_t* GetScriptName(CScript* _pScript);
-};
